Extract store address fields to a reusable component
